+++ src/plugins/bookmark-views/index.js
@@ -3,9 +3,11 @@
  * @copyright 2022, the Converse.js contributors
  * @license Mozilla Public License (MPLv2)
  */
+import './modals/bookmark-list.js';
+import './modals/bookmark-form.js';
 import '@converse/headless/plugins/muc/index.js';
-import BookmarkForm from './form.js';
-import BookmarksView from './bookmarks-list.js';
+import BookmarkForm from './components/bookmark-form.js';
+import BookmarksView from './components/bookmarks-list.js';
 import { _converse, api, converse } from '@converse/headless/core';
 import { bookmarkableChatRoomView } from './mixins.js';
 import { getHeadingButtons, removeBookmarkViaEvent, addBookmarkViaEvent } from './utils.js';