+++ CHANGES.md
@@ -12,6 +12,7 @@
 - Update `nick` attribute on ChatRoom when user nickname changes
 - Restrict editing of MUC messages to ones with the same XEP-0421 occupant ID
 - Remove unfurls for links removed in a subsequent message correction
+- Bookmarks now appear in a modal and the `hide_open_bookmarks` config setting has been removed.
 - #1004: Stop using fonts to render icons and use SVG instead
 - #2797: Weird unicode characters rendering outside of line-height
 - #2870: Fix for multiple URLs to be linkified when sent together in chat and adds a test for this.
@@ -21,6 +22,7 @@
 - #3005: Fix MUC messages with a fallback body not rendering.
 - #3007: Fix links becoming text when a message is edited
 - #3018: Fix MUC icons not functioning.
+
 
 ## 9.1.1 (2022-05-05)